+++ src/FluentURL.spec.js
@@ -1,7 +1,7 @@
 const FluentURL = require('./FluentURL.js');
 const FluentURLSearchParams = require('./FluentURLSearchParams.js');
 
-describe('FluentURL', () => {
+describe('FluentURL constructor', () => {
 	const testUrl = 'https://user:pass@sub.example.com:8443/foo/?q=hello#hash';
 
 	it('should construct with url string', () => {
@@ -27,6 +27,63 @@ describe('FluentURL', () => {
 		});
 		expect(url.toString()).toBe('https://example.com/foo?a=one');
 	});
+	it('should construct with no arguments', () => {
+		const url = new FluentURL();
+		url
+			.protocol('https')
+			.hostname('example.com')
+			.pathname('/foo')
+			.searchObject({ a: 'one' });
+		expect(url.toString()).toBe('https://example.com/foo?a=one');
+	});
+});
+describe('FluentURL methods', () => {
+	const testUrl = 'https://user:pass@sub.example.com:8443/foo/?q=hello#hash';
+	it('should export all details', () => {
+		const url = new FluentURL(testUrl);
+		expect(url.export()).toEqual({
+			hash: '#hash',
+			host: 'sub.example.com:8443',
+			hostname: 'sub.example.com',
+			href: testUrl,
+			isRelative: false,
+			origin: 'https://sub.example.com:8443',
+			password: 'pass',
+			pathname: '/foo/',
+			port: '8443',
+			protocol: 'https:',
+			search: '?q=hello',
+			searchObject: { q: 'hello' },
+			username: 'user',
+		});
+	});
+	it('should export only valid requested details', () => {
+		const url = new FluentURL(testUrl);
+		expect(url.export(['host', 'protocol', 'notAThing'])).toEqual({
+			host: 'sub.example.com:8443',
+			protocol: 'https:',
+		});
+	});
+	it('should import requested values', () => {
+		const url = new FluentURL();
+		url.import({
+			hash: '#hash',
+			hostname: 'sub.example.com',
+			password: 'pass',
+			pathname: '/foo/',
+			port: '8443',
+			protocol: 'https:',
+			searchObject: { q: 'hello' },
+			username: 'user',
+			notAThing: 42,
+		});
+		expect(url.href()).toBe(testUrl);
+	});
+	it('should clone ok', () => {
+		const url1 = new FluentURL(testUrl).searchObject({ a: 'one' });
+		const url2 = url1.clone();
+		expect(url1.href()).toBe(url2.href());
+	});
 	it('should get and set hash', () => {
 		const url = new FluentURL(testUrl);
 		expect(url.hash()).toBe('#hash');
@@ -48,8 +105,9 @@ describe('FluentURL', () => {
 	it('should get and set href', () => {
 		const url = new FluentURL(testUrl);
 		expect(url.href()).toBe(testUrl);
-		expect(url.href('http://httpbin.org/foo/?bar#baz')).toBe(url);
-		expect(url.href()).toBe('http://httpbin.org/foo/?bar#baz');
+		const newHref = 'http://httpbin.org/foo/?bar=1#baz';
+		expect(url.href(newHref)).toBe(url);
+		expect(url.href()).toBe(newHref);
 	});
 	it('should get origin', () => {
 		const url = new FluentURL(testUrl);
@@ -107,6 +165,15 @@ describe('FluentURL', () => {
 		const newParams = {};
 		expect(url.search(newParams)).toBe(url);
 		expect(url.search()).toBe('');
+	});
+	it('should get search as object', () => {
+		const url = new FluentURL('http://example.com/search?q=hello');
+		expect(url.searchObject()).toEqual({ q: 'hello' });
+	});
+	it('should set search from object', () => {
+		const url = new FluentURL('http://example.com/search?q=hello');
+		url.searchObject({ a: 'one', b: 'two' });
+		expect(url.search()).toEqual('?a=one&b=two');
 	});
 	it('should serialize with toString', () => {
 		const url = new FluentURL(testUrl);
@@ -235,6 +302,64 @@ describe('FluentURL search params methods', () => {
 	it('should support qsValues', () => {
 		const url = new FluentURL(`${testUrl}&a[]=1&a[]=2`);
 		expect(Array.from(url.qsValues())).toEqual(['hello', '1', '2']);
+	});
+	it('should support qsToString', () => {
+		const url = new FluentURL('https://example.com/foo/?a=1&b=2');
+		expect(url.qsToString()).toEqual('a=1&b=2');
+	});
+});
+describe('FluentURL relative URLs', () => {
+	const testUrl = '/foo/?q=hello#hash';
+
+	it('should return relative path', () => {
+		const url = new FluentURL(testUrl);
+		expect(url.href()).toBe('/foo/?q=hello#hash');
+		expect(url.isRelative()).toBe(true);
+	});
+	it('should return relative path with ../', () => {
+		const url = new FluentURL('../foo/?q=hello#hash');
+		expect(url.href()).toBe('../foo/?q=hello#hash');
+		expect(url.isRelative()).toBe(true);
+	});
+	it('should return relative path with ./', () => {
+		const url = new FluentURL('./foo/?q=hello#hash');
+		expect(url.href()).toBe('./foo/?q=hello#hash');
+		expect(url.isRelative()).toBe(true);
+	});
+	it('should return null for protocol', () => {
+		const url = new FluentURL(testUrl);
+		expect(url.protocol()).toBe(undefined);
+	});
+	it('should return null for port', () => {
+		const url = new FluentURL(testUrl);
+		expect(url.port()).toBe(undefined);
+	});
+	it('should return null for host', () => {
+		const url = new FluentURL(testUrl);
+		expect(url.host()).toBe(undefined);
+	});
+	it('should return null for hostname', () => {
+		const url = new FluentURL(testUrl);
+		expect(url.hostname()).toBe(undefined);
+	});
+	it('should return null for username', () => {
+		const url = new FluentURL(testUrl);
+		expect(url.username()).toBe(undefined);
+	});
+	it('should return null for password', () => {
+		const url = new FluentURL(testUrl);
+		expect(url.password()).toBe(undefined);
+	});
+	it('should return null for origin', () => {
+		const url = new FluentURL(testUrl);
+		expect(url.origin()).toBe(undefined);
+	});
+	it('should allowing making absolute URL relative', () => {
+		const url = new FluentURL(
+			'https://user:password@example.com:8443' + testUrl
+		);
+		url.makeRelative();
+		expect(url.href()).toBe(testUrl);
 	});
 });
 describe('FluentURL hash routing', () => {