+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 ## Change Log
 
+### v1.2.0 on 2022-09-11
+
+- Add support for relative URLs and empty URLs
+- New methods searchObject, import, export, clone, qsToString, qsToObject
+- New FluentURLSearchParams methods toObject
+- Allow FluentURLSearchParams#set to accept an object
+- Use new codecov uploader
+- Update npm dev dependencies
+
 ### v1.1.0 on 2022-05-12
 
 - Add support for hash routing with `hashPath()` and `hashSearch()`